+++ app/javascript/vue/components/radials/object/components/asserted_environments/asserted_environments_annotator.vue
@@ -12,6 +12,7 @@
       class="margin-medium-top"
       :list="list"
       label="uri_label"
+      annotator
       @delete="removeItem"
     />
   </div>